requirements.txt
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.0.0
oauth2client==4.1.3
psycopg2==2.9.6
requests==2.28.2
freezegun==1.2.2
responses==0.23.1
flake8==6.0.0